refactor: 자동차의 이름을 CarName이라는 클래스로 포장
