+++ src/main/java/racingcar/domain/car/Car.java
@@ -1,22 +1,17 @@
 package racingcar.domain.car;
 
+import racingcar.domain.car.vo.CarName;
+
 public class Car {
-    private final String name;
+    private final CarName name;
     private final CarEngine engine;
 
     private int position;
 
     public Car(String name, CarEngine engine) {
-        validateName(name);
-        this.name = name;
+        this.name = new CarName(name);
         this.engine = engine;
         this.position = 0;
-    }
-
-    private void validateName(String name) {
-        if (name.length() > 5) {
-            throw new IllegalArgumentException("이름은 5글자를 초과할 수 없습니다.");
-        }
     }
 
     public int getPosition() {
